rename staking api query files
