+++ packages/plugin-staking-api/src/index.tsx
@@ -4,14 +4,14 @@
 import { ApolloProvider } from '@apollo/client'
 
 export * from './Client'
-export * from './queries/useCanFastUnstake'
-export * from './queries/usePoolCandidates'
-export * from './queries/usePoolEraPoints'
-export * from './queries/usePoolRewards'
-export * from './queries/useRewards'
-export * from './queries/useTokenPrice'
-export * from './queries/useUnclaimedRewards'
-export * from './queries/useValidatorEraPoints'
+export * from './queries/canFastUnstake'
+export * from './queries/poolCandidates'
+export * from './queries/poolEraPoints'
+export * from './queries/poolRewards'
+export * from './queries/rewards'
+export * from './queries/tokenPrice'
+export * from './queries/unclaimedRewards'
+export * from './queries/validatorEraPoints'
 export * from './util'
 
 export { ApolloProvider }
